refactor: changes the export of filter as a type to a enum correctly
