+++ src/types.ts
@@ -15,20 +15,6 @@ import { SegmentationUniformConfig as SegmentationUniformConfigType, Segmentatio
  * Contains metadata about the file including its status, size, and video information
  */
 export type { File } from '../generated/common';
-/**
- * Filter operators for metadata, video info, and file property filtering
- * Used across different APIs for consistent filtering behavior
- */
-export enum FilterOperator {
-  NotEqual = "NotEqual",
-  Equal = "Equal",
-  LessThan = "LessThan",
-  GreaterThan = "GreaterThan",
-  ContainsAny = "ContainsAny",
-  ContainsAll = "ContainsAll",
-  In = "In",
-  Like = "Like"
-}
 
 
 /**
